Stop throwing an error if history cannot be deleted
This is the first time I've managed to get it to error out in the first place. In this case I had duplicated a form with Duplicate Post which broke something. The form was not able to be deleted permanently as it errored to this. The cause of the error was a missing table.

I'm leaving this hanging for now as there might be implications for changing the behaviour, and as stated, the error didn't even happen without using a 3rd party plugin.